py/requirements.txt
jsonschema-typed-v2>=0.8.0
